.github: workflows: Use zephyr-runner v2
This commit updates the GitHub Actions CI workflow to use the new
zephyr-runner v2.

As part of this, it also updates the workflow to use Buildah and Podman
to build the ci-repo-cache image because the zephyr-runner v2 has
stricter security requirements and do not support running privileged
pods, which is necessary for launching Docker daemon.

Signed-off-by: Stephanos Ioannidis <root@stephanos.io>