.github: workflows: Set TMPDIR to under work volume
This commit updates the CI workflow to set the TMPDIR environment
variable to `/__w/tmp`, which is under the runner work volume in order
to ensure that sufficient free space is available for the temporary
image layers to be stored.

Note that the TMPDIR environment variable is used by Podman/Buildah to
set `image_copy_tmp_dir`.

Signed-off-by: Stephanos Ioannidis <root@stephanos.io>